requirements.txt
# Core Dependencies
sqlalchemy>=2.0.0
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
python-dotenv
alembic
pydantic>=2.0.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.910
flake8>=4.0.0
